feat: refactor waitlist page and add new components for global stats and requests section
