+++ libs/burn-waitlist/src/lib/components/index.ts
@@ -9,3 +9,7 @@ export * from './wallet-connection-warning';
 export * from './network-warning';
 export * from './waitlist-balances';
 export * from './price-chart';
+export * from './global-stats';
+export * from './user-aggregates';
+export * from './waitlist-form-section';
+export * from './requests-section';
